Update app.py
- The mechanism for automatic deletion of temporary voice channels has been reworked
- Implemented caching of channel IDs created by the bot
- The function for checking and deleting empty temporary channels has been reworked
- By default, channels are deleted automatically every 60 seconds

+++ app.py
@@ -13,13 +13,14 @@ bot = commands.Bot(command_prefix="!", intents=intents)
 
 voice_channel_id = [123456789, 123456789]  # Replace with the desired voice channel ID
 category_id = [123456789, 123456789]  # Replace with the desired category ID
-
+created_channels = set()
 
 @bot.event
 async def on_ready():
     print("Boosty developer: https://boosty.to/mao-mao")
     print("GitHub: https://github.com/rinnyuwu")
     print("Donation Alerts: https://www.donationalerts.com/r/rinnyuwu")
+    check_empty_channels.start()
 
 @bot.event
 async def on_voice_state_update(member, before, after):
@@ -35,22 +36,21 @@ async def on_voice_state_update(member, before, after):
             name=f"{member.name}'s Channel", category=chosen_category
         )
 
+        created_channels.add(new_voice_channel.id)
+
         await member.move_to(new_voice_channel)
 
         await new_voice_channel.set_permissions(
             member, manage_channels=True, connect=True, speak=True
         )
 
-        await check_and_delete_channel(new_voice_channel)
-
-async def check_and_delete_channel(channel: disnake.VoiceChannel):
-    await asyncio.sleep(60)
-
-    if channel.guild.get_channel(channel.id) is None:
-        return
-
-    if len(channel.members) == 0:
-        await channel.delete()
+@tasks.loop(seconds=60)
+async def check_empty_channels():
+    for channel_id in list(created_channels):
+        channel = bot.get_channel(channel_id)
+        if channel and len(channel.members) == 0:
+            await channel.delete()
+            created_channels.remove(channel_id)
 
 bot.run("INSERT-TOKEN") # To run the bot, you'll need a token which you get by creating an application in the Discord Developer Portal (https://discord.com/developers/)
                         # Replace INSERT-TOKEN with your token here
